Reorganised the flow for how data from the Overpass API is retrieved and processed
