+++ src/scripts/map/search.ts
@@ -1,5 +1,5 @@
 import { MessageBoxError } from "../messages.js";
-import { Overpass } from "../overpass.js";
+import { overpassSearch } from "../overpass/index.js";
 import { getElement } from "../supplement/elements.js";
 
 const SEARCH_FORM = getElement("search-form", HTMLFormElement);
@@ -49,7 +49,7 @@ async function search(searchTerm: string) {
 	setSearching(true);
 
 	try {
-		await Overpass.search(searchTerm);
+		await overpassSearch(searchTerm);
 	} catch (error) {
 		if (error instanceof MessageBoxError) error.display();
 		else console.error(error);
